# database/account.go
package database

import "github.com/jinzhu/gorm"

type Account struct {
	gorm.Model            //leaving this ananymous field here so gorm:embedded tag isn't necessary
	AccountHolder string  `json:"accountHolder" binding:"required"`
	AccountType   string  `json:"accountType" binding:"required"`
	Balance       float64 `json:"balance" binding:"required"`
}

type AccountService interface {
	Service[Account]
}