# docs/swagger.yaml
definitions:
  database.Account:
    properties:
      accountHolder:
        type: string
      accountType:
        type: string
      balance:
        type: number
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    required:
    - accountHolder
    - accountType
    - balance
    type: object
  database.Transaction:
    properties:
      account:
        $ref: '#/definitions/database.Account'
      accountID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      transactionAmount:
        type: number
      transactionType:
        type: string
      updatedAt:
        type: string
    required:
    - accountID
    - transactionAmount
    - transactionType
    type: object
  routes.AuthToken:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
info:
  contact: {}
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: lists all account records in the DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Account'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: list all account records
      tags:
      - Accounts
    post:
      consumes:
      - application/json
      description: creates an account record in the DB
      parameters:
      - description: create account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/database.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Account'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: create an account record
      tags:
      - Accounts
  /accounts/:id:
    delete:
      consumes:
      - application/json
      description: deletes an account record from the DB
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: delete an account record
      tags:
      - Accounts
    get:
      consumes:
      - application/json
      description: fetches an account record by id from the DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Account'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: fetches an account record by id
      tags:
      - Accounts
    put:
      consumes:
      - application/json
      description: updates an account record in the DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Account'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: update an account record
      tags:
      - Accounts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Generates a JWT Token for use with Authorized endpoints
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.AuthToken'
        "401":
          description: Unauthorized
      summary: Login generates a JWT token
      tags:
      - Authentication
  /health:
    get:
      consumes:
      - application/json
      description: Returns a message if the service is working correctly
      produces:
      - application/json
      responses:
        "200":
          description: The service is working.
          schema:
            type: string
      summary: Server Health
      tags:
      - Health
  /transactions:
    get:
      consumes:
      - application/json
      description: lists all transaction records in the DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Transaction'
            type: array
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: list all transaction records
      tags:
      - Tranasctions
    post:
      consumes:
      - application/json
      description: allows a transaction to be created in the database if the account
        exists
      parameters:
      - description: Create Transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/database.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: create a transaction record
      tags:
      - Tranasctions
  /transactions/:id:
    delete:
      consumes:
      - application/json
      description: allows a transaction to be deleted from the database
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: delete a transaction record
      tags:
      - Tranasctions
    get:
      consumes:
      - application/json
      description: allows a transaction to be deleted from the database
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/database.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: delete a transaction record
      tags:
      - Tranasctions
    put:
      consumes:
      - application/json
      description: update the amount of a transaction record
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Transaction'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: update the amount of a transaction record
      tags:
      - Tranasctions
swagger: "2.0"